This change improves update and query backward reads (by ~65-93%) and backward writes (by ~45-71%). The overall time improvement across all heap benchmarks is ~45%:

```
= cf8b2a46ffd..8e2a8b39b5c: Embedders Heap total time: 139448 ms (improved by -44.9%)
  - update/wasm64_update_read_bwd_1gb_step_4kb_checkpoint time improved by -93.7%
  - update/wasm32_update_read_bwd_1gb_step_4kb_checkpoint time improved by -93.6%
  - query/wasm64_query_read_bwd_1gb_step_4kb_checkpoint time improved by -91.4%
  - query/wasm32_query_read_bwd_1gb_step_4kb_checkpoint time improved by -91.2%
  - query/wasm32_query_read_bwd_1gb_step_4kb_new_allocation time improved by -89.5%
  - query/wasm64_query_read_bwd_1gb_step_4kb_new_allocation time improved by -89.5%
  - update/wasm32_update_read_bwd_1gb_step_4kb_new_allocation time improved by -89.4%
  - update/wasm64_update_read_bwd_1gb_step_4kb_new_allocation time improved by -89.4%
  - query/wasm32_query_read_bwd_1gb_step_16kb_new_allocation time improved by -87.2%
  - query/wasm64_query_read_bwd_1gb_step_16kb_new_allocation time improved by -87.2%
  - update/wasm32_update_read_bwd_1gb_step_16kb_new_allocation time improved by -86.6%
  - update/wasm64_update_read_bwd_1gb_step_16kb_new_allocation time improved by -86.6%
  - update/wasm32_update_read_bwd_1gb_step_8_checkpoint time improved by -86.3%
  - query/wasm32_query_read_bwd_1gb_step_8_checkpoint time improved by -85.9%
  - update/wasm64_update_read_bwd_1gb_step_8_checkpoint time improved by -85.5%
  - query/wasm32_query_read_bwd_1gb_step_8_new_allocation time improved by -84.8%
  - query/wasm64_query_read_bwd_1gb_step_8_checkpoint time improved by -84.6%
  - update/wasm32_update_read_bwd_1gb_step_8_new_allocation time improved by -84.6%
  - query/wasm64_query_read_bwd_1gb_step_8_new_allocation time improved by -84.1%
  - update/wasm64_update_read_bwd_1gb_step_8_new_allocation time improved by -83.9%
  - update/wasm32_update_read_bwd_1gb_step_16kb_checkpoint time improved by -81.8%
  - update/wasm64_update_read_bwd_1gb_step_16kb_checkpoint time improved by -80.9%
  - update/wasm32_update_read_bwd_1gb_step_4kb_page_delta time improved by -78.2%
  - update/wasm64_update_read_bwd_1gb_step_4kb_page_delta time improved by -77.7%
  - update/wasm32_update_read_bwd_1gb_step_8_page_delta time improved by -75.8%
  - update/wasm64_update_read_bwd_1gb_step_8_page_delta time improved by -75.4%
  - query/wasm32_query_read_bwd_1gb_step_16kb_checkpoint time improved by -74.7%
  - query/wasm64_query_read_bwd_1gb_step_16kb_checkpoint time improved by -74%
  - query/wasm64_query_write_bwd_1gb_step_4kb_new_allocation time improved by -71.8%
  - query/wasm32_query_write_bwd_1gb_step_4kb_new_allocation time improved by -71.5%
  - query/wasm64_query_write_bwd_1gb_step_16kb_new_allocation time improved by -71.3%
  - query/wasm64_query_write_bwd_1gb_step_4kb_page_delta time improved by -69.4%
  - query/wasm32_query_write_bwd_1gb_step_4kb_page_delta time improved by -68.6%
  - query/wasm64_query_read_bwd_1gb_step_4kb_page_delta time improved by -68.3%
  - query/wasm32_query_write_bwd_1gb_step_16kb_new_allocation time improved by -68.1%
  - query/wasm32_query_write_bwd_1gb_step_8_new_allocation time improved by -67.8%
  - query/wasm32_query_read_bwd_1gb_step_4kb_page_delta time improved by -67.7%
  - query/wasm64_query_write_bwd_1gb_step_8_new_allocation time improved by -67.3%
  - query/wasm64_query_write_bwd_1gb_step_8_page_delta time improved by -66.3%
  - query/wasm64_query_write_bwd_1gb_step_4kb_checkpoint time improved by -65.7%
  - query/wasm32_query_read_bwd_1gb_step_8_page_delta time improved by -65.6%
  - query/wasm32_query_write_bwd_1gb_step_8_page_delta time improved by -65.1%
  - query/wasm64_query_read_bwd_1gb_step_8_page_delta time improved by -65.1%
  - query/wasm32_query_write_bwd_1gb_step_4kb_checkpoint time improved by -65%
  - query/wasm32_query_write_bwd_1gb_step_8_checkpoint time improved by -63.6%
  - query/wasm64_query_write_bwd_1gb_step_8_checkpoint time improved by -62.9%
  - query/wasm32_query_write_bwd_1gb_step_16kb_checkpoint time improved by -62.8%
  - query/wasm64_query_write_bwd_1gb_step_16kb_checkpoint time improved by -62.4%
  - update/wasm32_update_write_bwd_1gb_step_4kb_new_allocation time improved by -51%
  - update/wasm64_update_write_bwd_1gb_step_4kb_new_allocation time improved by -50.8%
  - update/wasm64_update_write_bwd_1gb_step_8_new_allocation time improved by -49.4%
  - update/wasm32_update_write_bwd_1gb_step_8_new_allocation time improved by -48.9%
  - update/wasm64_update_write_bwd_1gb_step_4kb_checkpoint time improved by -48.5%
  - update/wasm64_update_write_bwd_1gb_step_4kb_page_delta time improved by -47.7%
  - update/wasm32_update_write_bwd_1gb_step_4kb_checkpoint time improved by -47.6%
  - update/wasm32_update_write_bwd_1gb_step_8_checkpoint time improved by -47.4%
  - update/wasm32_update_write_bwd_1gb_step_4kb_page_delta time improved by -47.3%
  - update/wasm64_update_write_bwd_1gb_step_8_page_delta time improved by -46.3%
  - update/wasm64_update_write_bwd_1gb_step_8_checkpoint time improved by -45.8%
  - update/wasm32_update_write_bwd_1gb_step_8_page_delta time improved by -45.3%
  ```

There are four above 5% regressions in forward reads/writes:

```
wasm64_query_write_fwd_1gb_step_4kb_new_allocation: 767 ms -> 844 ms (regressed by 9.9%)
wasm32_query_write_fwd_1gb_step_8_checkpoint: 1134 ms -> 1225 ms (regressed by 8%)
wasm32_query_write_fwd_1gb_step_16kb_checkpoint: 282 ms -> 303 ms (regressed by 7.2%)
wasm64_query_read_fwd_1gb_step_4kb_checkpoint: 244 ms -> 260 ms (regressed by 6.6%)
```

Having also forward improvements above 5%:
```
  - query/wasm64_query_read_fwd_1gb_step_8_checkpoint time improved by -10.3%
  - update/wasm32_update_read_fwd_1gb_step_8_checkpoint time improved by -8%
  - query/wasm64_query_write_fwd_1gb_step_8_page_delta time improved by -6.5%
```
Likely those forward improvements and regressions are just noise.